requirements.txt
a2a-sdk[http-server]>=0.3.0
openai>=1.0.0
pydantic>=2.0.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
click>=8.0.0
starlette>=0.27.0
httpx>=0.24.0
sse-starlette>=0.1.0